added c_api for flint module, currently untested
