+++ src/flint/numpy_flint.c
@@ -29,15 +29,8 @@
 
 #include "flint.h"
 
-/// @brief A python flint object
-/// @param obval The internal c representation of the flint object
-typedef struct {
-    PyObject_HEAD
-    flint obval;
-} PyFlint;
-
-/// @brief The flint PyTypeObject
-static PyTypeObject PyFlint_Type;
+#define NUMPY_FLINT_MODULE
+#include "numpy_flint.h"
 
 /// @brief The array of data members of the flint object
 PyMemberDef pyflint_members[] = {
@@ -55,24 +48,6 @@ PyMemberDef pyflint_members[] = {
 // ###############################
 // This section contains some convenience functions that are used by the method
 // implementations below.
-
-/// @brief Check if an object is a flint
-/// @param ob The PyObject to check
-/// @return 1 if the object is a flint, 0 otherwise
-static inline int PyFlint_Check(PyObject* ob) {
-    return PyObject_IsInstance(ob, (PyObject*) &PyFlint_Type);
-}
-
-/// @brief Create a PyFlint object from a c flint struct.
-/// @param f The c flint struct
-/// @return A new PyFlint object that contains a copy of f
-static inline PyObject* PyFlint_FromFlint(flint f) {
-    PyFlint* p = (PyFlint*) PyFlint_Type.tp_alloc(&PyFlint_Type, 0);
-    if (p) {
-        p->obval = f;
-    }
-    return (PyObject*) p;
-}
 
 /// @brief A macro to check if an object is a PyFlint and grab the c struct.
 /// @param f The c flint struct to copy the data into
@@ -1335,9 +1310,10 @@ PyMODINIT_FUNC PyInit_flint(void) {
     PyObject* numpy;
     PyObject* numpy_dict;
     PyArray_Descr* npyflint_descr;
-    int NPY_FLINT;
     PyArray_Descr* from_descr;
     int arg_types[3];
+    static void* PyFlint_API[2];
+    PyObject* c_api_object;
     // Create the new module
     m = PyModule_Create(&moduledef);
     if (m==NULL) {
@@ -1549,6 +1525,17 @@ PyMODINIT_FUNC PyInit_flint(void) {
         Py_DECREF(m);
         PyErr_Print();
         PyErr_SetString(PyExc_SystemError, "Could not add flint.flint type to module flint.");
+        return NULL;
+    }
+    // Register PyFlint_Type and NPY_FLINT with the c api
+    PyFlint_API[0] = (void*) &PyFlint_Type;
+    PyFlint_API[1] = (void*) &NPY_FLINT;
+    c_api_object = PyCapsule_New((void*) PyFlint_API, "flint._C_API", NULL);
+    if (PyModule_AddObject(m, "_C_API", c_api_object) < 0) {
+        Py_XDECREF(c_api_object);
+        Py_DECREF(m);
+        PyErr_Print();
+        PyErr_SetString(PyExc_SystemError, "Could not add flint._C_API to flint module.");
         return NULL;
     }
